When a job fails without junit failures, capture the tail of build-log.txt
Will capture some failure modes that ci-operator can't capture.

Support a new index mode for build-log and show the selector now that
we have two types.
